requirements.txt
Flask==1.0.2
splitio_client[uwsgi,cpphash]==8.4.1
uwsgi
six
